refactor(backends): move the Firecracker API client and VMM process into fc/
First slice of the Firecracker substrate: the four modules with no
orchestration coupling at all. fc_api, fork_namespace and lifecycle had
no crate:: or super:: references; daemon reached only for shell and ui,
both already in mvm-vmm::host.

The per-VM path layout (abs_vms_dir, resolve_running_vm_dir, fc_pid_path,
firecracker_vsock_uds_path) goes with them: daemon reads it, and leaving
it behind would have meant mvm-backends calling back into mvm-runtime for
its own directory names.

Visibility widens to pub only where mvm-runtime still names an item
across the crate line — pub(crate) cannot cross it, and these have real
callers in microvm/ until the rest of the tree follows.

mvm-runtime::microvm re-exports every moved name at its original path, so
crate::microvm::<name> and mvm_runtime::microvm::<name> resolve unchanged.
