+++ crates/mvm-backends/Cargo.toml
@@ -20,6 +20,7 @@ mvm-core = { workspace = true }
 mvm-net = { workspace = true }
 mvm-agentd = { workspace = true }
 anyhow = { workspace = true }
+tempfile = { workspace = true }
 tracing = { workspace = true }
 libkrun-sys = { workspace = true }
 serde = { workspace = true }
