+++ crates/mvm-backends/src/fc/lifecycle.rs
